test: Rename "router_newbenchmark" to "router_benchmark".
Since, the original "router_benchmark" has been deleted the "new" prefix is
no longer necessary.
