+++ acceptance/router_benchmark/brload/BUILD.bazel
@@ -4,10 +4,10 @@ load("//:scion.bzl", "scion_go_binary")
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "github.com/scionproto/scion/acceptance/router_newbenchmark/brload",
+    importpath = "github.com/scionproto/scion/acceptance/router_benchmark/brload",
     visibility = ["//visibility:private"],
     deps = [
-        "//acceptance/router_newbenchmark/cases:go_default_library",
+        "//acceptance/router_benchmark/cases:go_default_library",
         "//pkg/log:go_default_library",
         "//pkg/private/serrors:go_default_library",
         "//pkg/scrypto:go_default_library",
