+++ acceptance/router_benchmark/cases/out.go
@@ -26,24 +26,24 @@ import (
 	"github.com/scionproto/scion/pkg/slayers/path/scion"
 )
 
-// Topology (see accept/router_newbenchmark/conf/topology.json)
+// Topology (see accept/router_benchmark/conf/topology.json)
 //    AS2 (br2) ---+== (br1a) AS1 (br1b) ---- (br4) AS4
 //                 |
 //    AS3 (br3) ---+
 // See topo.go
 
-// oneIn generates one packet of incoming traffic into AS1 at br1a. The outcome is a raw packet
+// oneOut generates one packet of outgoing traffic from AS1 at br1a. The outcome is a raw packet
 // that the test must feed into the router.
-func oneIn(payload string, mac hash.Hash, flowId uint32) []byte {
+func oneOut(payload string, mac hash.Hash, flowId uint32) []byte {
 
 	var (
-		originIA       = ISDAS(2)
-		originIP       = PublicIP(2, 1)
+		originIA       = ISDAS(1)
+		originIP       = InternalIP(1, 2)
 		originHost     = HostAddr(originIP)
-		srcIP, srcPort = PublicIPPort(2, 1)
-		dstIP, dstPort = PublicIPPort(1, 2)
-		targetIA       = ISDAS(1)
-		targetIP       = InternalIP(1, 2)
+		srcIP, srcPort = InternalIPPort(1, 2)
+		dstIP, dstPort = InternalIPPort(1, 1)
+		targetIA       = ISDAS(2)
+		targetIP       = PublicIP(2, 1)
 		targetHost     = HostAddr(targetIP)
 	)
 
@@ -58,7 +58,7 @@ func oneIn(payload string, mac hash.Hash, flowId uint32) []byte {
 	sp := &scion.Decoded{
 		Base: scion.Base{
 			PathMeta: scion.MetaHdr{
-				CurrHF: 1,
+				CurrHF: 0,
 				SegLen: [3]uint8{2, 0, 0},
 			},
 			NumINF:  1,
@@ -68,20 +68,21 @@ func oneIn(payload string, mac hash.Hash, flowId uint32) []byte {
 			{
 				SegID:     0x111,
 				Timestamp: util.TimeToSecs(time.Now()),
-				ConsDir:   false,
+				ConsDir:   true,
 			},
 		},
 		HopFields: []path.HopField{
-			{ConsIngress: 1, ConsEgress: 0}, // From there (non-consdir)
-			{ConsIngress: 0, ConsEgress: 2}, // <- Processed here (non-consdir)
+			{ConsIngress: 0, ConsEgress: 2}, // <- Processed here
+			{ConsIngress: 1, ConsEgress: 0}, // Going there
 		},
 	}
 
 	// Calculate MACs...
-	// Seg0: Hops are in non-consdir.
-	sp.HopFields[1].Mac = path.MAC(mac, sp.InfoFields[0], sp.HopFields[1], nil)
-	sp.InfoFields[0].UpdateSegID(sp.HopFields[1].Mac)
-	sp.HopFields[0].Mac = path.MAC(FakeMAC(2), sp.InfoFields[0], sp.HopFields[0], nil)
+	// Seg0: Hops are in consdir.
+	sp.HopFields[0].Mac = path.MAC(mac, sp.InfoFields[0], sp.HopFields[0], nil)
+	sp.InfoFields[0].UpdateSegID(sp.HopFields[0].Mac) // tmp
+	sp.HopFields[1].Mac = path.MAC(FakeMAC(2), sp.InfoFields[0], sp.HopFields[0], nil)
+	sp.InfoFields[0].SegID = 0x111 // Restore to initial.
 
 	// End-to-end. Src is the originator and Dst is the final destination.
 	scionL := &slayers.SCION{
@@ -115,17 +116,18 @@ func oneIn(payload string, mac hash.Hash, flowId uint32) []byte {
 	); err != nil {
 		panic(err)
 	}
+
 	return input.Bytes()
 }
 
-// In generates numDistinct packets (each with a unique flowID) with the given payload
-// constructed to cause "in" traffic at the br1a router.
+// Out generates numDistinct packets (each with a unique flowID) with the given payload
+// constructed to cause "out" traffic at the br1a router.
 // numDistrinct is a small number, only to enable multiple parallel streams. Each distinct packet
 // is meant to be replayed a large number of times for performance measurement.
-func In(payload string, mac hash.Hash, numDistinct int) (string, string, [][]byte) {
+func Out(payload string, mac hash.Hash, numDistinct int) (string, string, [][]byte) {
 	packets := make([][]byte, numDistinct)
 	for i := 0; i < numDistinct; i++ {
-		packets[i] = oneIn(payload, mac, uint32(i+1))
+		packets[i] = oneOut(payload, mac, uint32(i+1))
 	}
-	return DeviceName(1, 2), DeviceName(1, 0), packets
+	return DeviceName(1, 0), DeviceName(1, 2), packets
 }
